requirements.txt
Django==1.6.2
WebOb==1.3.1
WebTest==2.0.14
argparse==1.2.1
beautifulsoup4==4.3.2
boto==2.25.0
coverage==3.7.1
dj-database-url==0.2.2
dj-static==0.0.5
django-bootstrap3==2.5.6
django-storages==1.1.8
django-tastypie==0.11.0
django-tenancy==0.1.2
django-toolbelt==0.0.1
django-webtest==1.7.6
docutils==0.11
gunicorn==18.0
psycopg2==2.5.2
pystache==0.5.3
python-dateutil==2.2
python-mimeparse==0.1.4
pytz==2013.9
six==1.5.2
static==1.0.2
waitress==0.8.8
wsgiref==0.1.2
